refactor(app): 成员列表点任何一行都在第三列开资料，不再进会话草稿
上一版只改了外部成员的点击。按要求扩到全部：成员、智能体、外部联系人点击一律
在第三列开资料 tab（target `actor/<id>`），右键「查看资料」走同一个地方——列表
是用来找人的，进会话由资料页里的「开始会话」按钮显式触发。

为了不让资料出现两套实现，把 ActorDetailDialog 的正文抽成
ActorDetailContent（头像/在线态/详情表/客户端版本/重新邀请/移除，以及底部动作
栏），两个宿主共用：

- ActorDetailDialog：只剩 modal 外壳，仍服务左侧 recents 和本机 daemon 行那些
  没有第三列可用的入口；传 onClose，所以「关闭」按钮和「开始会话/移除后自动关
  闭」的行为不变。
- ActorDetailPane（新，components/main-content/）：从目录里按 id 取行，取不到就
  单独 fetch（tab 被恢复、或该 actor 已被移除时至少能显示真实状态）；tab 由标签
  栏关闭，所以不传 onClose，底部不再有多余的关闭按钮。多一个 extraSections 插槽，
  用来挂「参与的会话」列表——对所有 kind 都有用，放在 pane 而不是共用正文里，
  dialog 的观感保持原样。

原来的 ExternalActorDetail 删掉，它的字段并进共用正文：外部联系人多「来源」
「来源 ID」两行、没有「开始会话」（改成一句说明），并且不再显示恒为「—」的角色
行；created_at 那行对外部联系人叫「首次出现」，对成员仍叫「加入时间」。

验证：packages/app tsc --noEmit 通过；改动文件 eslint 干净；vitest 全量
446 files / 2882 tests 全绿（新增「成员点击开 tab 且不进草稿」用例，
i18n-parity 的死键守卫也是这一轮 firstSeen 复用后才重新变绿）。
